feature_builder/requirements.txt
pandas==2.1.1
sqlalchemy==2.0.21
psycopg2-binary==2.9.9
numpy==1.26.0
python-dotenv==1.0.0
